added a command handler
